feat: move sources to modal
